Move metric/fig calcs to spd/core_metrics_and_figs.py
